+++ anytrans-20250707/alibabacloud_anytrans20250707/models.py
@@ -591,6 +591,593 @@ class BatchTranslateResponse(TeaModel):
         return self
 
 
+class BatchTranslateForHtmlRequestExtConfig(TeaModel):
+    def __init__(
+        self,
+        skip_csi_check: bool = None,
+    ):
+        self.skip_csi_check = skip_csi_check
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.skip_csi_check is not None:
+            result['skipCsiCheck'] = self.skip_csi_check
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('skipCsiCheck') is not None:
+            self.skip_csi_check = m.get('skipCsiCheck')
+        return self
+
+
+class BatchTranslateForHtmlRequestExtExamples(TeaModel):
+    def __init__(
+        self,
+        src: str = None,
+        tgt: str = None,
+    ):
+        self.src = src
+        self.tgt = tgt
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.src is not None:
+            result['src'] = self.src
+        if self.tgt is not None:
+            result['tgt'] = self.tgt
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('src') is not None:
+            self.src = m.get('src')
+        if m.get('tgt') is not None:
+            self.tgt = m.get('tgt')
+        return self
+
+
+class BatchTranslateForHtmlRequestExtTerminologies(TeaModel):
+    def __init__(
+        self,
+        src: str = None,
+        tgt: str = None,
+    ):
+        self.src = src
+        self.tgt = tgt
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.src is not None:
+            result['src'] = self.src
+        if self.tgt is not None:
+            result['tgt'] = self.tgt
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('src') is not None:
+            self.src = m.get('src')
+        if m.get('tgt') is not None:
+            self.tgt = m.get('tgt')
+        return self
+
+
+class BatchTranslateForHtmlRequestExtTextTransform(TeaModel):
+    def __init__(
+        self,
+        to_lower: bool = None,
+        to_title: bool = None,
+        to_upper: bool = None,
+    ):
+        self.to_lower = to_lower
+        self.to_title = to_title
+        self.to_upper = to_upper
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.to_lower is not None:
+            result['toLower'] = self.to_lower
+        if self.to_title is not None:
+            result['toTitle'] = self.to_title
+        if self.to_upper is not None:
+            result['toUpper'] = self.to_upper
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('toLower') is not None:
+            self.to_lower = m.get('toLower')
+        if m.get('toTitle') is not None:
+            self.to_title = m.get('toTitle')
+        if m.get('toUpper') is not None:
+            self.to_upper = m.get('toUpper')
+        return self
+
+
+class BatchTranslateForHtmlRequestExt(TeaModel):
+    def __init__(
+        self,
+        config: BatchTranslateForHtmlRequestExtConfig = None,
+        domain_hint: str = None,
+        examples: List[BatchTranslateForHtmlRequestExtExamples] = None,
+        sensitives: List[str] = None,
+        terminologies: List[BatchTranslateForHtmlRequestExtTerminologies] = None,
+        text_transform: BatchTranslateForHtmlRequestExtTextTransform = None,
+    ):
+        self.config = config
+        self.domain_hint = domain_hint
+        self.examples = examples
+        self.sensitives = sensitives
+        self.terminologies = terminologies
+        self.text_transform = text_transform
+
+    def validate(self):
+        if self.config:
+            self.config.validate()
+        if self.examples:
+            for k in self.examples:
+                if k:
+                    k.validate()
+        if self.terminologies:
+            for k in self.terminologies:
+                if k:
+                    k.validate()
+        if self.text_transform:
+            self.text_transform.validate()
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.config is not None:
+            result['config'] = self.config.to_map()
+        if self.domain_hint is not None:
+            result['domainHint'] = self.domain_hint
+        result['examples'] = []
+        if self.examples is not None:
+            for k in self.examples:
+                result['examples'].append(k.to_map() if k else None)
+        if self.sensitives is not None:
+            result['sensitives'] = self.sensitives
+        result['terminologies'] = []
+        if self.terminologies is not None:
+            for k in self.terminologies:
+                result['terminologies'].append(k.to_map() if k else None)
+        if self.text_transform is not None:
+            result['textTransform'] = self.text_transform.to_map()
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('config') is not None:
+            temp_model = BatchTranslateForHtmlRequestExtConfig()
+            self.config = temp_model.from_map(m['config'])
+        if m.get('domainHint') is not None:
+            self.domain_hint = m.get('domainHint')
+        self.examples = []
+        if m.get('examples') is not None:
+            for k in m.get('examples'):
+                temp_model = BatchTranslateForHtmlRequestExtExamples()
+                self.examples.append(temp_model.from_map(k))
+        if m.get('sensitives') is not None:
+            self.sensitives = m.get('sensitives')
+        self.terminologies = []
+        if m.get('terminologies') is not None:
+            for k in m.get('terminologies'):
+                temp_model = BatchTranslateForHtmlRequestExtTerminologies()
+                self.terminologies.append(temp_model.from_map(k))
+        if m.get('textTransform') is not None:
+            temp_model = BatchTranslateForHtmlRequestExtTextTransform()
+            self.text_transform = temp_model.from_map(m['textTransform'])
+        return self
+
+
+class BatchTranslateForHtmlRequest(TeaModel):
+    def __init__(
+        self,
+        app_name: str = None,
+        ext: BatchTranslateForHtmlRequestExt = None,
+        format: str = None,
+        scene: str = None,
+        source_language: str = None,
+        target_language: str = None,
+        text: Dict[str, Any] = None,
+        workspace_id: str = None,
+    ):
+        self.app_name = app_name
+        self.ext = ext
+        self.format = format
+        self.scene = scene
+        # This parameter is required.
+        self.source_language = source_language
+        # This parameter is required.
+        self.target_language = target_language
+        # This parameter is required.
+        self.text = text
+        # This parameter is required.
+        self.workspace_id = workspace_id
+
+    def validate(self):
+        if self.ext:
+            self.ext.validate()
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.app_name is not None:
+            result['appName'] = self.app_name
+        if self.ext is not None:
+            result['ext'] = self.ext.to_map()
+        if self.format is not None:
+            result['format'] = self.format
+        if self.scene is not None:
+            result['scene'] = self.scene
+        if self.source_language is not None:
+            result['sourceLanguage'] = self.source_language
+        if self.target_language is not None:
+            result['targetLanguage'] = self.target_language
+        if self.text is not None:
+            result['text'] = self.text
+        if self.workspace_id is not None:
+            result['workspaceId'] = self.workspace_id
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('appName') is not None:
+            self.app_name = m.get('appName')
+        if m.get('ext') is not None:
+            temp_model = BatchTranslateForHtmlRequestExt()
+            self.ext = temp_model.from_map(m['ext'])
+        if m.get('format') is not None:
+            self.format = m.get('format')
+        if m.get('scene') is not None:
+            self.scene = m.get('scene')
+        if m.get('sourceLanguage') is not None:
+            self.source_language = m.get('sourceLanguage')
+        if m.get('targetLanguage') is not None:
+            self.target_language = m.get('targetLanguage')
+        if m.get('text') is not None:
+            self.text = m.get('text')
+        if m.get('workspaceId') is not None:
+            self.workspace_id = m.get('workspaceId')
+        return self
+
+
+class BatchTranslateForHtmlShrinkRequest(TeaModel):
+    def __init__(
+        self,
+        app_name: str = None,
+        ext_shrink: str = None,
+        format: str = None,
+        scene: str = None,
+        source_language: str = None,
+        target_language: str = None,
+        text_shrink: str = None,
+        workspace_id: str = None,
+    ):
+        self.app_name = app_name
+        self.ext_shrink = ext_shrink
+        self.format = format
+        self.scene = scene
+        # This parameter is required.
+        self.source_language = source_language
+        # This parameter is required.
+        self.target_language = target_language
+        # This parameter is required.
+        self.text_shrink = text_shrink
+        # This parameter is required.
+        self.workspace_id = workspace_id
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.app_name is not None:
+            result['appName'] = self.app_name
+        if self.ext_shrink is not None:
+            result['ext'] = self.ext_shrink
+        if self.format is not None:
+            result['format'] = self.format
+        if self.scene is not None:
+            result['scene'] = self.scene
+        if self.source_language is not None:
+            result['sourceLanguage'] = self.source_language
+        if self.target_language is not None:
+            result['targetLanguage'] = self.target_language
+        if self.text_shrink is not None:
+            result['text'] = self.text_shrink
+        if self.workspace_id is not None:
+            result['workspaceId'] = self.workspace_id
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('appName') is not None:
+            self.app_name = m.get('appName')
+        if m.get('ext') is not None:
+            self.ext_shrink = m.get('ext')
+        if m.get('format') is not None:
+            self.format = m.get('format')
+        if m.get('scene') is not None:
+            self.scene = m.get('scene')
+        if m.get('sourceLanguage') is not None:
+            self.source_language = m.get('sourceLanguage')
+        if m.get('targetLanguage') is not None:
+            self.target_language = m.get('targetLanguage')
+        if m.get('text') is not None:
+            self.text_shrink = m.get('text')
+        if m.get('workspaceId') is not None:
+            self.workspace_id = m.get('workspaceId')
+        return self
+
+
+class BatchTranslateForHtmlResponseBodyDataTranslationListUsage(TeaModel):
+    def __init__(
+        self,
+        input_tokens: int = None,
+        output_tokens: int = None,
+        total_tokens: int = None,
+    ):
+        self.input_tokens = input_tokens
+        self.output_tokens = output_tokens
+        self.total_tokens = total_tokens
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.input_tokens is not None:
+            result['inputTokens'] = self.input_tokens
+        if self.output_tokens is not None:
+            result['outputTokens'] = self.output_tokens
+        if self.total_tokens is not None:
+            result['totalTokens'] = self.total_tokens
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('inputTokens') is not None:
+            self.input_tokens = m.get('inputTokens')
+        if m.get('outputTokens') is not None:
+            self.output_tokens = m.get('outputTokens')
+        if m.get('totalTokens') is not None:
+            self.total_tokens = m.get('totalTokens')
+        return self
+
+
+class BatchTranslateForHtmlResponseBodyDataTranslationList(TeaModel):
+    def __init__(
+        self,
+        code: int = None,
+        index: str = None,
+        message: str = None,
+        translation: str = None,
+        usage: BatchTranslateForHtmlResponseBodyDataTranslationListUsage = None,
+    ):
+        self.code = code
+        self.index = index
+        self.message = message
+        self.translation = translation
+        self.usage = usage
+
+    def validate(self):
+        if self.usage:
+            self.usage.validate()
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.code is not None:
+            result['code'] = self.code
+        if self.index is not None:
+            result['index'] = self.index
+        if self.message is not None:
+            result['message'] = self.message
+        if self.translation is not None:
+            result['translation'] = self.translation
+        if self.usage is not None:
+            result['usage'] = self.usage.to_map()
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('code') is not None:
+            self.code = m.get('code')
+        if m.get('index') is not None:
+            self.index = m.get('index')
+        if m.get('message') is not None:
+            self.message = m.get('message')
+        if m.get('translation') is not None:
+            self.translation = m.get('translation')
+        if m.get('usage') is not None:
+            temp_model = BatchTranslateForHtmlResponseBodyDataTranslationListUsage()
+            self.usage = temp_model.from_map(m['usage'])
+        return self
+
+
+class BatchTranslateForHtmlResponseBodyData(TeaModel):
+    def __init__(
+        self,
+        translation_list: List[BatchTranslateForHtmlResponseBodyDataTranslationList] = None,
+    ):
+        self.translation_list = translation_list
+
+    def validate(self):
+        if self.translation_list:
+            for k in self.translation_list:
+                if k:
+                    k.validate()
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        result['translationList'] = []
+        if self.translation_list is not None:
+            for k in self.translation_list:
+                result['translationList'].append(k.to_map() if k else None)
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        self.translation_list = []
+        if m.get('translationList') is not None:
+            for k in m.get('translationList'):
+                temp_model = BatchTranslateForHtmlResponseBodyDataTranslationList()
+                self.translation_list.append(temp_model.from_map(k))
+        return self
+
+
+class BatchTranslateForHtmlResponseBody(TeaModel):
+    def __init__(
+        self,
+        code: str = None,
+        data: BatchTranslateForHtmlResponseBodyData = None,
+        http_status_code: str = None,
+        message: str = None,
+        request_id: str = None,
+        success: bool = None,
+    ):
+        self.code = code
+        self.data = data
+        self.http_status_code = http_status_code
+        self.message = message
+        self.request_id = request_id
+        self.success = success
+
+    def validate(self):
+        if self.data:
+            self.data.validate()
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.code is not None:
+            result['code'] = self.code
+        if self.data is not None:
+            result['data'] = self.data.to_map()
+        if self.http_status_code is not None:
+            result['httpStatusCode'] = self.http_status_code
+        if self.message is not None:
+            result['message'] = self.message
+        if self.request_id is not None:
+            result['requestId'] = self.request_id
+        if self.success is not None:
+            result['success'] = self.success
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('code') is not None:
+            self.code = m.get('code')
+        if m.get('data') is not None:
+            temp_model = BatchTranslateForHtmlResponseBodyData()
+            self.data = temp_model.from_map(m['data'])
+        if m.get('httpStatusCode') is not None:
+            self.http_status_code = m.get('httpStatusCode')
+        if m.get('message') is not None:
+            self.message = m.get('message')
+        if m.get('requestId') is not None:
+            self.request_id = m.get('requestId')
+        if m.get('success') is not None:
+            self.success = m.get('success')
+        return self
+
+
+class BatchTranslateForHtmlResponse(TeaModel):
+    def __init__(
+        self,
+        headers: Dict[str, str] = None,
+        status_code: int = None,
+        body: BatchTranslateForHtmlResponseBody = None,
+    ):
+        self.headers = headers
+        self.status_code = status_code
+        self.body = body
+
+    def validate(self):
+        if self.body:
+            self.body.validate()
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.headers is not None:
+            result['headers'] = self.headers
+        if self.status_code is not None:
+            result['statusCode'] = self.status_code
+        if self.body is not None:
+            result['body'] = self.body.to_map()
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('headers') is not None:
+            self.headers = m.get('headers')
+        if m.get('statusCode') is not None:
+            self.status_code = m.get('statusCode')
+        if m.get('body') is not None:
+            temp_model = BatchTranslateForHtmlResponseBody()
+            self.body = temp_model.from_map(m['body'])
+        return self
+
+
 class GetDocTranslateTaskRequest(TeaModel):
     def __init__(
         self,
@@ -2218,8 +2805,10 @@ class SubmitDocTranslateTaskResponse(TeaModel):
 class SubmitHtmlTranslateTaskRequestExtConfig(TeaModel):
     def __init__(
         self,
+        callback_url: str = None,
         skip_csi_check: bool = None,
     ):
+        self.callback_url = callback_url
         self.skip_csi_check = skip_csi_check
 
     def validate(self):
@@ -2231,12 +2820,16 @@ class SubmitHtmlTranslateTaskRequestExtConfig(TeaModel):
             return _map
 
         result = dict()
+        if self.callback_url is not None:
+            result['callbackUrl'] = self.callback_url
         if self.skip_csi_check is not None:
             result['skipCsiCheck'] = self.skip_csi_check
         return result
 
     def from_map(self, m: dict = None):
         m = m or dict()
+        if m.get('callbackUrl') is not None:
+            self.callback_url = m.get('callbackUrl')
         if m.get('skipCsiCheck') is not None:
             self.skip_csi_check = m.get('skipCsiCheck')
         return self
@@ -4368,6 +4961,7 @@ class TextTranslateRequestExtTextTransform(TeaModel):
 class TextTranslateRequestExt(TeaModel):
     def __init__(
         self,
+        agent: str = None,
         config: TextTranslateRequestExtConfig = None,
         domain_hint: str = None,
         examples: List[TextTranslateRequestExtExamples] = None,
@@ -4375,6 +4969,7 @@ class TextTranslateRequestExt(TeaModel):
         terminologies: List[TextTranslateRequestExtTerminologies] = None,
         text_transform: TextTranslateRequestExtTextTransform = None,
     ):
+        self.agent = agent
         self.config = config
         self.domain_hint = domain_hint
         self.examples = examples
@@ -4402,6 +4997,8 @@ class TextTranslateRequestExt(TeaModel):
             return _map
 
         result = dict()
+        if self.agent is not None:
+            result['agent'] = self.agent
         if self.config is not None:
             result['config'] = self.config.to_map()
         if self.domain_hint is not None:
@@ -4422,6 +5019,8 @@ class TextTranslateRequestExt(TeaModel):
 
     def from_map(self, m: dict = None):
         m = m or dict()
+        if m.get('agent') is not None:
+            self.agent = m.get('agent')
         if m.get('config') is not None:
             temp_model = TextTranslateRequestExtConfig()
             self.config = temp_model.from_map(m['config'])